move add_button script to its own file in _ext
